requirements.txt
google-api-python-client>=2.85.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
nltk>=3.8.1
requests>=2.28.2
python-dotenv>=1.0.0
pathlib>=1.0.1